+++ common-libs/common/src/dto/shared.dto.ts
@@ -1,6 +1,10 @@
 import { EventName, EventComponent, VerificationMetadata } from '@affinidi/affinity-metrics-lib'
 import { DidResolver } from '../shared/DidResolver'
 
+export type DocumentLoader = (
+  iri: string,
+) => Promise<{ contextUrl: string | null; document: Record<string, any>; documentUrl: string | null } | undefined>
+
 export class AffinityOptions {
   apiKey?: string
   registryUrl?: string
@@ -10,6 +14,8 @@ export class AffinityOptions {
   useCache?: boolean
   cacheMaxSize?: number
   cacheTtlInMin?: number
+  resolveLegacyElemLocally?: boolean
+  beforeDocumentLoader?: DocumentLoader
 }
 
 export class EventOptions {