+++ sdk/core/src/CommonNetworkMember/BaseNetworkMember.ts
@@ -137,9 +137,15 @@ export abstract class BaseNetworkMember {
       didAuthAdapter,
     })
     this._holderService = new HolderService(
-      { registryUrl, metricsUrl, accessApiKey },
+      {
+        registryUrl,
+        metricsUrl,
+        accessApiKey,
+      },
       platformCryptographyTools,
       eventComponent,
+      options.otherOptions?.resolveLocallyElemMethod,
+      options.otherOptions?.beforeDocumentLoader,
     )
     this._affinity = new Affinity(
       {
@@ -147,6 +153,8 @@ export abstract class BaseNetworkMember {
         registryUrl: registryUrl,
         metricsUrl: metricsUrl,
         component: eventComponent,
+        resolveLegacyElemLocally: options.otherOptions?.resolveLocallyElemMethod,
+        beforeDocumentLoader: options.otherOptions?.beforeDocumentLoader,
       },
       platformCryptographyTools,
     )