+++ sdk/core/CHANGELOG.md
@@ -1,3 +1,11 @@
+# release 7.11.0 (2022-12-19)
+* added an option for HolderService to resolve legacy elem locally
+# release 7.10.1 (2022-12-09)
+updated `getCredentials` to filter corrupted credentials without throwing exception
+# release 7.10.0 (2022-12-07)
+* added an option to pass custom before hook for document loader
+# release 7.9.0 (2022-12-07)
+* added a flag to resolve legacy elem locally
 # release 7.8.0 (2022-11-29)
 * add security fix for signIn passwordless flow
 # release 7.7.0 (2022-11-29)